Test.QuickCheck.Classes.Monoid.Monus

Description

This module provides Laws definitions for classes exported by Data.Monoid.Monus.

Documentation

monusLaws :: forall a. (Arbitrary a, Show a, Eq a, Monus a) => Proxy a -> Laws

Laws for instances of Monus.

Tests the following laws:

a <\> a == mempty

mempty <\> a == mempty

a <> (b <\> a) == b <> (a <\> b)

(a <\> b) <\> c == a <\> (b <> c)

a <\> b == stripPrefixOverlap b a

a <\> b == stripSuffixOverlap b a

Note that the following superclass laws are not included:

commutativeLaws

monoidLaws

overlappingGCDMonoidLaws

overlappingGCDMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, OverlappingGCDMonoid a) => Proxy a -> Laws

Laws for instances of OverlappingGCDMonoid.

Tests the following laws:

overlap a b <> stripPrefixOverlap a b == b

stripSuffixOverlap b a <> overlap a b == a

stripOverlap a b & \(_, x, _) -> x == overlap a b

stripOverlap a b & \(_, _, x) -> x == stripPrefixOverlap a b

stripOverlap a b & \(x, _, _) -> x == stripSuffixOverlap b a

Note that the following superclass laws are not included:

monoidLaws

leftReductiveLaws

rightReductiveLaws