Test.QuickCheck.Classes.Monoid.Null

Description

This module provides Laws definitions for classes exported by Data.Monoid.Null.

Null

monoidNullLaws :: forall a. (Arbitrary a, Show a, Eq a, MonoidNull a) => Proxy a -> Laws

Laws for instances of MonoidNull.

Tests the following law:

null a == (a == mempty)

Note that the following superclass laws are not included:

monoidLaws

Positive

positiveMonoidLaws :: forall a. (Arbitrary a, Show a, Eq a, PositiveMonoid a) => Proxy a -> Laws

Laws for instances of PositiveMonoid.

Tests the following law:

null (a <> b) == (null a && null b)

Note that the following superclass laws are not included:

monoidNullLaws