Test.QuickCheck.Classes.Semigroup

Description

This module provides:

Support for testing Semigroup subclass instances with QuickCheck.

Reusable properties in the form of Laws definitions.

In general:

Each Laws definition includes properties that relate to just one class.

Laws for superclasses are not transitively included.

Therefore, when testing laws for a particular class, you should make sure to also test laws for all superclasses.

Monus

monusLaws :: forall a. (Arbitrary a, Show a, Eq a, Monus a) => Proxy a -> Laws

Laws for instances of Monus.

Tests the following properties:

a <\> b == stripPrefixOverlap b a

a <\> b == stripSuffixOverlap b a

Note that the following superclass laws are not included:

commutativeLaws

overlappingGCDMonoidLaws

Group

groupLaws :: forall a. (Arbitrary a, Show a, Eq a, Group a) => Proxy a -> Laws

Laws for instances of Group.

Tests the following properties:

invert mempty == mempty

a <> invert a == mempty

invert a <> a == mempty

a ~~ mempty == a

a ~~ a == mempty

a ~~ b == a <> invert b

n >= 0 ==> pow a n == mconcat (replicate n a)

n <= 0 ==> pow a n == invert (mconcat (replicate (abs n) a))