Test.QuickCheck.Classes.Semigroup.Factorial

Description

This module provides Laws definitions for classes exported by Data.Semigroup.Factorial.

Documentation

factorialLaws :: forall a. (Arbitrary a, Show a, Eq a, Factorial a) => Proxy a -> Laws

Laws for instances of Factorial.

Includes the following laws:

length a == Data.List.length (factors a)

maybe a sconcat (nonEmpty           $ factors a) ==         a
maybe a sconcat (nonEmpty $ reverse $ factors a) == reverse a

all (\f -> factors f == [f]) (factors a)

primePrefix a == foldr (\x _ -> x) a a
primeSuffix a == foldl (\_ x -> x) a a

foldl  f x a == Data.List.foldl  f x (factors a)
foldl' f x a == Data.List.foldl' f x (factors a)
foldr  f x a == Data.List.foldr  f x (factors a)

Note that the following superclass laws are not included:

semigroupLaws