Test.QuickCheck.Classes.Semigroup.Cancellative

Description

Commutative

commutativeLaws :: forall a. (Arbitrary a, Show a, Eq a, Commutative a) => Proxy a -> Laws

Laws for instances of Commutative.

Tests the following property:

a <> b == b <> a

Cancellative

cancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, Cancellative a) => Proxy a -> Laws

Laws for instances of Cancellative.

Tests the following properties:

(a <> b) </> a == Just b

(a <> b) </> b == Just a

Note that the following superclass laws are not included:

leftCancellativeLaws

rightCancellativeLaws

reductiveLaws

leftCancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, LeftCancellative a) => Proxy a -> Laws

Laws for instances of LeftCancellative.

Tests the following property:

stripPrefix a (a <> b) == Just b

Note that the following superclass laws are not included:

leftReductiveLaws

rightCancellativeLaws :: forall a. (Arbitrary a, Show a, Eq a, RightCancellative a) => Proxy a -> Laws

Laws for instances of RightCancellative.

Tests the following property:

stripSuffix b (a <> b) == Just a

Note that the following superclass laws are not included:

rightReductiveLaws

Reductive

reductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, Reductive a) => Proxy a -> Laws

Laws for instances of Reductive.

Tests the following properties:

a </> b == stripPrefix b a

a </> b == stripSuffix b a

maybe a (b <>) (a </> b) == a

maybe a (<> b) (a </> b) == a

Note that the following superclass laws are not included:

commutativeLaws

leftReductiveLaws

rightReductiveLaws

leftReductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, LeftReductive a) => Proxy a -> Laws

Laws for instances of LeftReductive.

Tests the following properties:

a `isPrefixOf` (a <> b)

isPrefixOf a b == isJust (stripPrefix a b)

maybe b (a <>) (stripPrefix a b) == b

rightReductiveLaws :: forall a. (Arbitrary a, Show a, Eq a, RightReductive a) => Proxy a -> Laws

Laws for instances of RightReductive.

Tests the following properties:

b `isSuffixOf` (a <> b)

isSuffixOf a b == isJust (stripSuffix a b)

maybe b (<> a) (stripSuffix a b) == b